"Add chainID to method signatures and update encode config in tests"
This commit introduces a chainID argument to various method signatures in the gravity keeper. These changes were necessary to ensure that we're calling the correct chain when interacting with the keeper. There was also a couple of updates to the `MakeTestEncodingConfig()` function in the test files to match new function signatures. Meanwhile, some error checking routines (hooks.go) were modified to return error rather than void. Lastly, we had to correct import path in governance_proposals.go due to a version upgrade in cosmos-sdk's gov module.