"Refactor client error response handling in gravity module"
This commit refactors the handling of error responses in certain areas of the "gravity" module of the Cosmos-SDK. Rather than relying on the rest.WriteErrorResponse and rest.PostProcessResponse methods from the Cosmos-SDK's REST package, we've introduced equivalent functions directly in our module. These new methods were introduced in query.go file.

This allows us to have more control over the error response behavior directly within our module, and potentially make modifications that may not be fit for contribution back to the larger Cosmos-SDK project. It also adds the functionality to wrap responses with the height, providing more detailed information to the client. A notable addition here is the encoding/json package, used to process JSON messages in responses.

Remember this makes part of our efforts to ensure orderly and accurate error reporting, crucial for effective debugging and user feedback.