Remove IBCMetadataProposal implementation
The IBCMetadataProposal implementation was removed from the RegisteredImplementations in codec.go and the methods related to it removed from governance_proposals.go as it is no longer necessary. The genesis.go has been updated to include an array of GravityChain structures in place of the removed implementation to enhance flexibility and scalability.