Add gravitytypes to app Modules and ChainIds parameter
The gravitytypes module name was moved in modules order in 'app/app.go' for better arrangement. Also, a new parameter 'ChainIds' was added to the 'x/gravity/types/genesis.go'. This addition allows for more flexibility in interacting with different blockchain networks. The function 'validateChainIds' was added to check the type of the incoming chain ids, aiming to prevent type mismatches.