"Refactor chain handling in Gravity module"
This commit includes significant changes in the Gravity module handling of chain data and actions. Previously, chain information was passed around in the form of an array of strings, making the code harder to read and follow. This has been changed by creating a new message, ChainParam, which has everything necessary to define a chain - ID, addresses, and contract details. On the concept of chains, these changes update the gravity keeper to take in a chainID parameter for its function calls. This adjustment means that functions are now performed on a specific chain instead of assuming a single global chain. Furthermore, chain information isn't simply accessed from the chains field in Params but is processed and validated when needed. It's quite a substantial refactoring aimed at making the code  cleaner and easier to understand, improve its effectiveness, and set things up for better handling of multiple chains.