Update testnet configuration for crossfi
The configuration for the testnet environment has been updated to switch from the 'mineplex-chaind' binary to 'crossfi', and associated properties have been adjusted accordingly. This includes changes to the seed address, enabling JSON-RPC with specific APIs, and updating the genesis parameters. Also, the accounts' coins have been increased by 100000000000000000000000000mpx and 100000000000000000000000000000xfi. This was necessary to reflect the recent changes and settings for the crossfi testnet environment.